brique39/requirements-pro.txt
# Core Data & ML
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
lightgbm==3.3.5
joblib==1.2.0
pyarrow==12.0.1

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Kafka
kafka-python==2.0.2

# ML Tools
shap==0.42.1
mlflow==2.7.1

# Monitoring
prometheus-client==0.17.1